The following changes were implemented:
- **Staff Management & Creation**:
    - Implemented a new `CreateStaffUseCase` and integrated it into `UserCreateViewModel`.
    - Enhanced `UserCreateScreen` with dynamic role filtering based on the current user's permissions.
    - Added region and district selection logic for creating `ROLE_DISTRICT_ADMIN` users.
    - Introduced `CreateStaffUserDto` to handle staff-specific registration details including roles and district assignments.
- **UI & UX Enhancements**:
    - Migrated `TextField` components to `OutlinedTextField` across `UserCreateScreen`, `CoachCreateScreen`, and `SettingsScreen` for a more modern look.
    - Added password visibility toggles to all password input fields using `VisualTransformation` and `IconButton`.
    - Standardized input field styling with `RoundedCornerShape(16.dp)` and consistent border colors.
    - Updated navigation in `CoachesScreen` to redirect to the unified `UserCreateTypeScreen`.
- **Data Precision Updates**:
    - Refactored revenue and growth percentage fields from `Int` to `Double` across DTOs and Domain models (`WeeklyReport`, `StadiumRevenue`) to support decimal values.
    - Updated `ReportsScreen` and `AdminRepositoryImpl` to handle the updated numeric types.
- **Authentication & Session Improvements**:
    - Consolidated logout logic in `SessionManager` and `AuthRepositoryImpl` to ensure consistent state clearing.
    - Improved `HttpClientFactory` to trigger an automatic logout when receiving a 401 response with `TOKEN_INVALID` or `REFRESH_TOKEN_NOT_FOUND` error bodies.
- **Refactoring**:
    - Cleaned up unused imports and simplified specialty selection in `CoachCreateScreen`.
    - Updated `AppModule` to include new use case and view model dependencies.

+++ composeApp/src/commonMain/kotlin/uz/coder/foottopbusiness/data/network/dto/admin/StadiumRevenue.kt
@@ -9,5 +9,5 @@ data class StadiumRevenue(
     @SerialName("stadiumId")
     val stadiumId: Int? = null,
     @SerialName("totalRevenue")
-    val totalRevenue: Int? = null
+    val totalRevenue: Double? = null
 )
