+++ composeApp/src/commonMain/kotlin/uz/coder/foottopbusiness/domain/model/StadiumRevenue.kt
@@ -2,5 +2,5 @@ package uz.coder.foottopbusiness.domain.model
 
 data class StadiumRevenue(
     val stadiumId: Int,
-    val totalRevenue: Int
+    val totalRevenue: Double
 )
