+++ composeApp/src/commonMain/kotlin/uz/coder/foottopbusiness/domain/model/WeeklyReport.kt
@@ -1,11 +1,11 @@
 package uz.coder.foottopbusiness.domain.model
 
 data class WeeklyReport(
-    val bookingsGrowthPercent: Int,
+    val bookingsGrowthPercent: Double,
     val dailyRevenue: List<DailyRevenue>,
     val previousWeekBookings: Int,
     val totalBookings: Int,
     val weekEnd: String,
     val weekStart: String,
-    val weeklyRevenue: Int
+    val weeklyRevenue: Double
 )
